fdm_sentinel/requirements.txt
fastapi
uvicorn[standard]
pywebpush
apscheduler
python-dotenv
torch
torchvision
pillow
opencv-python
torch
python-multipart
sse-starlette